PR_variables (#336)

+++ listings/ch02-common-programming-concepts/no_listing_05_mut_cant_change_type/src/lib.cairo
@@ -2,7 +2,7 @@
 use debug::PrintTrait;
 use traits::Into;
 fn main() {
-    let mut x: u32 = 2;
+    let mut x: u64 = 2;
     x.print();
     x = 100_felt252;
     x.print()
